use cosf sinf instead of cos sin to avoid conversion warning

+++ src/main.cxx
@@ -208,8 +208,8 @@ int main(void)
     for (int i = 0; i < 2 * n_spikes; ++i) {
         float phi_i = phi_0 + i * dPhi;
         float radius = radii[i % 2];
-        starPositions.push_back(radius * cos(phi_i));
-        starPositions.push_back(radius * sin(phi_i));
+        starPositions.push_back(radius * cosf(phi_i));
+        starPositions.push_back(radius * sinf(phi_i));
     }
     starPositions.push_back(0.f); // center point x
     starPositions.push_back(0.f); // center point y
